PDFBOX-3441: use long instead of int to handle huge files, as suggested by Michael Doswald
git-svn-id: https://svn.apache.org/repos/asf/pdfbox/branches/2.0@1754431 13f79535-47bb-0310-9956-ffa450edef68

+++ pdfbox/src/main/java/org/apache/pdfbox/pdfparser/COSParser.java
@@ -324,7 +324,7 @@ public class COSParser extends BaseParser
 
         COSDictionary dict = parseCOSDictionary();
         COSStream xrefStream = parseCOSStream(dict);
-        parseXrefStream(xrefStream, (int) objByteOffset, isStandalone);
+        parseXrefStream(xrefStream, objByteOffset, isStandalone);
         xrefStream.close();
 
         return dict.getLong(COSName.PREV);
@@ -2002,7 +2002,7 @@ public class COSParser extends BaseParser
                 {
                     try
                     {
-                        int currOffset = Integer.parseInt(splitString[0]);
+                        long currOffset = Long.parseLong(splitString[0]);
                         int currGenID = Integer.parseInt(splitString[1]);
                         COSObjectKey objKey = new COSObjectKey(currObjID, currGenID);
                         xrefTrailerResolver.setXRef(objKey, currOffset);
